requirements.txt
asttokens==2.2.1
backcall==0.2.0
binance-connector==3.1.1        
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12      
colorama==0.4.6
comm==0.1.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.31.2
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0     
google-auth-oauthlib==0.7.1     
googleapis-common-protos==1.57.0
greenlet==1.1.2
httplib2==0.21.0
idna==3.3
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.1
matplotlib-inline==0.1.6        
nest-asyncio==1.5.6
numpy==1.22.3
oauthlib==3.2.2
openpyxl==3.0.9
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
pipreqs==0.4.11
platformdirs==3.5.1
portion==2.2.0
prompt-toolkit==3.0.38
protobuf==4.21.9
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.18.0
Pygments==2.15.1
pyparsing==3.0.7
pyTelegramBotAPI==4.5.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==306
pyzmq==25.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.6
websocket-client==1.6.1
XlsxWriter==3.0.3
yarg==0.1.9
python_binance-1.0.19